// swift-deliver-backend/src/types/types.ts
export enum UserRole {
    RESTAURANT = "restaurant",
    DELIVERY = "delivery",
    CUSTOMER = "customer",
    ADMIN = "admin"
}

export enum OrderStatus {
    pending = "pending",
    confirmed = "confirmed",
    preparing = "preparing",
    picked_up = "picked_up",
    delivered = "delivered",
    cancelled = "cancelled",
    ready = "ready",
}

export enum JobName {
    ORDER_CONFIRMATION = "order-confirmation",
}